refactor(cli): extract cargo-build helpers into streamlib-cargo-build (#992)
Move the cargo-build orchestration logic out of streamlib-cli/src/commands/pack.rs
into a focused libs/streamlib-cargo-build crate so cargo xtask build-plugins
(#991) can reuse the same battle-tested cdylib-discovery / target-triple-staging
logic without depending on the whole CLI surface or duplicating it.

Extracted helpers (behavior unchanged, copied verbatim):

- host_target_triple() — reads STREAMLIB_CARGO_BUILD_HOST_TARGET from the
  crate's own build.rs (decoupled from streamlib-engine's identical helper)
- host_dylib_extension() — so/dylib/dll picker
- read_cargo_package_name() — extracts [package].name from Cargo.toml
- run_cargo_build_release() — invokes cargo --message-format=json
- parse_cargo_artifact_for_cdylib() — picks the host cdylib out of the
  JSON artifact stream (handles Cargo's dash-vs-underscore normalization)
- collect_host_dylibs_in_lib() — populated-lib branch enumeration

New helpers (for #991's xtask):

- discover_package_dirs() — walks root dirs looking for streamlib.yaml
- read_minimal_project_config() — parses streamlib.yaml into the
  ProjectConfigMinimal subset (package + processors only)
- has_rust_runtime_processors() — predicate for whether to cargo build
- discover_rust_impl_packages() — composed filter the xtask iterates over
- stage_built_cdylib() — copies a built cdylib into
  <package_dir>/lib/<host_triple>/<filename> so Runner::load_project
  can resolve it via the triple-keyed convention

Verification:

- 15 new unit tests in streamlib-cargo-build (host-triple shape, dylib
  ext, JSON parser, cargo-toml parser, package discovery, staging).
- 17 existing streamlib-cli tests still pass, including the four pack
  flow tests that exercise the populated-lib / no-build error /
  schemas-only / patch-rejection branches end-to-end via the extracted
  helpers.
- Schemas-only .slpkg packed before vs. after the extraction is
  byte-identical (SHA256-per-zip-entry comparison against
  packages/core). Wire format preserved.
- Workspace baseline: 1732 passed; 2 pre-existing polyglot-manual-source
  failures and 1 pre-existing vulkan-jpeg test compile error are
  unrelated to this change (reproduce on plain main).

The new crate's only deps are anyhow, serde, serde_json, serde_yaml,
toml, and streamlib-processor-schema (for ProjectConfigMinimal) —
zero engine deps, so the xtask in #991 can pull it in without dragging
the runtime / RHI / IPC layers along.

Closes #992.
