+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "libs/streamlib-engine",    # Engine internals (RHI, IPC, surface-share, runtime executor) — private dep of the SDK
     "libs/streamlib-surface-client", # Wire-format helpers for the per-runtime surface-sharing socket (Linux)
     "libs/streamlib-cli",       # Command-line interface for StreamLib runtime
+    "libs/streamlib-cargo-build", # Cargo-build orchestration helpers shared between `streamlib pack` and `cargo xtask build-plugins`
     "libs/streamlib-runtime",   # Standalone runtime binary (spawned by CLI)
     "libs/streamlib-processor-schema", # Processor schema types + YAML parser shared between streamlib runtime and streamlib-macros
     "libs/streamlib-jtd-codegen", # JTD-codegen pipeline: schema YAML → typed Rust/Python/TypeScript bindings
