+++ libs/streamlib-cargo-build/Cargo.toml
@@ -1,0 +1,26 @@
+# Copyright (c) 2025 Jonathan Fontanez
+# SPDX-License-Identifier: BUSL-1.1
+
+[package]
+name = "streamlib-cargo-build"
+description = "Cargo-build orchestration helpers shared between `streamlib pack` and `cargo xtask build-plugins` — host target triple probe, cdylib JSON-artifact parser, package-directory discovery, per-package staging"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license-file.workspace = true
+repository.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde_yaml = { workspace = true }
+toml = { workspace = true }
+tracing = { workspace = true }
+streamlib-processor-schema = { path = "../streamlib-processor-schema" }
+
+[dev-dependencies]
+tempfile = "3.14"
+
+[lints]
+workspace = true
