+++ libs/streamlib-cli/Cargo.toml
@@ -25,6 +25,11 @@ streamlib-processor-schema = { path = "../streamlib-processor-schema" }
 # `pkg install` / `pkg remove` — typed PackageRef, Manifest schema, etc.
 streamlib-idents = { path = "../streamlib-idents" }
 
+# Cargo-build orchestration helpers — `streamlib pack`'s auto-build branch
+# routes through this crate so the same logic is reused by
+# `cargo xtask build-plugins` without dragging in the full CLI surface.
+streamlib-cargo-build = { path = "../streamlib-cargo-build" }
+
 # JTD-codegen pipeline (drives `streamlib generate`)
 streamlib-jtd-codegen = { path = "../streamlib-jtd-codegen" }
