+++ libs/streamlib-cli/src/commands/pack.rs
@@ -3,15 +3,17 @@
 
 use std::fs::File;
 use std::io::{Read, Write};
-use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
+use std::path::Path;
 
 use anyhow::{Context, Result};
 use zip::write::FileOptions;
 use zip::ZipWriter;
 
 use streamlib::engine_internal::core::ProjectConfig;
-use streamlib::sdk::runtime::host_target_triple;
+use streamlib_cargo_build::{
+    collect_host_dylibs_in_lib, host_dylib_extension, host_target_triple,
+    read_cargo_package_name, run_cargo_build_release,
+};
 use streamlib_idents::Manifest;
 
 /// Pack a processor package into a .slpkg bundle.
@@ -259,206 +261,6 @@ pub fn pack(package_dir: &Path, output: Option<&Path>, no_build: bool) -> Result
     Ok(())
 }
 
-/// The dylib extension for the current host OS, matching Cargo's
-/// `cdylib` output convention (`.so` on Linux, `.dylib` on macOS,
-/// `.dll` on Windows).
-fn host_dylib_extension() -> &'static str {
-    if cfg!(target_os = "macos") {
-        "dylib"
-    } else if cfg!(target_os = "windows") {
-        "dll"
-    } else {
-        "so"
-    }
-}
-
-/// Enumerate dylibs in `<package_dir>/lib/` whose extension matches the
-/// host OS. Returns an empty Vec when the directory does not exist or
-/// contains no matching files — the caller decides whether that's an
-/// error (the auto-build path turns it into a `cargo build` invocation).
-fn collect_host_dylibs_in_lib(lib_dir: &Path, dylib_ext: &str) -> Result<Vec<PathBuf>> {
-    if !lib_dir.is_dir() {
-        return Ok(Vec::new());
-    }
-    let mut found = Vec::new();
-    for entry in std::fs::read_dir(lib_dir)
-        .with_context(|| format!("Failed to read lib/ directory: {}", lib_dir.display()))?
-    {
-        let entry = entry?;
-        let path = entry.path();
-        if path.extension().is_some_and(|ext| ext == dylib_ext) {
-            found.push(path);
-        }
-    }
-    found.sort();
-    Ok(found)
-}
-
-/// Read the package's `Cargo.toml` and return its `[package].name`. This
-/// is the value `cargo build -p <name>` accepts; it is **not** the same
-/// string as `streamlib.yaml`'s `package.name` (the two existing plugin
-/// examples deliberately use different names — see
-/// `examples/camera-rust-plugin/plugin/`).
-fn read_cargo_package_name(package_dir: &Path) -> Result<String> {
-    let cargo_toml_path = package_dir.join("Cargo.toml");
-    let body = std::fs::read_to_string(&cargo_toml_path).with_context(|| {
-        format!(
-            "Failed to read {} — auto-build requires a Cargo.toml \
-             alongside streamlib.yaml so cargo can locate the crate",
-            cargo_toml_path.display()
-        )
-    })?;
-    let parsed: toml::Value = toml::from_str(&body)
-        .with_context(|| format!("Failed to parse {}", cargo_toml_path.display()))?;
-    let name = parsed
-        .get("package")
-        .and_then(|p| p.get("name"))
-        .and_then(|n| n.as_str())
-        .ok_or_else(|| {
-            anyhow::anyhow!(
-                "{} has no [package].name — auto-build needs a named Cargo \
-                 crate to invoke `cargo build -p <name>`",
-                cargo_toml_path.display()
-            )
-        })?;
-    Ok(name.to_string())
-}
-
-/// Invoke `cargo build --release -p <cargo_name> --message-format=json`
-/// from `package_dir` and parse the JSON output for the produced
-/// host-OS cdylib path.
-///
-/// `--message-format=json` is the canonical way to discover Cargo
-/// artifact paths — it survives `CARGO_TARGET_DIR` overrides, workspace
-/// `[build].target-dir` config, custom `[profile]` settings, and
-/// anything else that would invalidate a hardcoded
-/// `<workspace>/target/release/<file>` assumption.
-///
-/// Cargo's progress output (the `Compiling foo …` lines and compiler
-/// diagnostics) is left inherited on stderr so a cold build does not
-/// appear hung. Only stdout — the JSON message stream — is captured.
-fn run_cargo_build_release(
-    package_dir: &Path,
-    cargo_name: &str,
-    dylib_ext: &str,
-) -> Result<PathBuf> {
-    // `--features <name>/plugin` enables the `export_plugin!` invocation
-    // in the package's `lib.rs`. The feature is default-off so force-link
-    // rlib consumers don't pull in the unmangled `STREAMLIB_PLUGIN`
-    // symbol (multiple rlibs with the same static collide at link time);
-    // pack flips it on so the produced cdylib carries the symbol the
-    // runtime dlopens.
-    let features_flag = format!("{}/plugin", cargo_name);
-    println!(
-        "Building {} (cargo build --release -p {} --features {})",
-        cargo_name, cargo_name, features_flag
-    );
-    let output = Command::new("cargo")
-        .arg("build")
-        .arg("--release")
-        .arg("--message-format=json")
-        .arg("-p")
-        .arg(cargo_name)
-        .arg("--features")
-        .arg(&features_flag)
-        .current_dir(package_dir)
-        .stdout(Stdio::piped())
-        .stderr(Stdio::inherit())
-        .output()
-        .with_context(|| {
-            format!(
-                "Failed to invoke `cargo build --release -p {} --features {}` in {}",
-                cargo_name,
-                features_flag,
-                package_dir.display()
-            )
-        })?;
-
-    if !output.status.success() {
-        // Cargo's own error output already streamed to the user's
-        // terminal via inherited stderr; the bail message just names
-        // the operation that failed.
-        anyhow::bail!(
-            "cargo build --release -p {} failed (run from {}). \
-             See cargo's output above.",
-            cargo_name,
-            package_dir.display(),
-        );
-    }
-
-    let stdout = String::from_utf8(output.stdout).with_context(|| {
-        format!(
-            "cargo build output for {} was not valid UTF-8",
-            cargo_name
-        )
-    })?;
-
-    parse_cargo_artifact_for_cdylib(&stdout, cargo_name, dylib_ext)?.ok_or_else(|| {
-        anyhow::anyhow!(
-            "cargo build --release -p {} completed but produced no \
-             host-OS cdylib (`*.{}`). Confirm the crate declares \
-             `crate-type = [\"cdylib\"]` in [lib].",
-            cargo_name,
-            dylib_ext
-        )
-    })
-}
-
-/// Scan one stream of `--message-format=json` cargo output for the host
-/// cdylib artifact belonging to `cargo_name`. Returns the absolute path
-/// of the matching dylib if any `compiler-artifact` message lists a
-/// cdylib produced for the named crate.
-///
-/// Cargo normalizes crate-target names by replacing dashes with
-/// underscores (so the package `grayscale-plugin` builds the cdylib
-/// target `grayscale_plugin`). The comparison accepts both forms.
-fn parse_cargo_artifact_for_cdylib(
-    cargo_json_output: &str,
-    cargo_name: &str,
-    dylib_ext: &str,
-) -> Result<Option<PathBuf>> {
-    let dot_ext = format!(".{}", dylib_ext);
-    let normalized = cargo_name.replace('-', "_");
-    for line in cargo_json_output.lines() {
-        let line = line.trim();
-        if line.is_empty() || !line.starts_with('{') {
-            continue;
-        }
-        let msg: serde_json::Value = match serde_json::from_str(line) {
-            Ok(v) => v,
-            Err(_) => continue, // Non-JSON lines (rare with --message-format=json) are skipped.
-        };
-        if msg.get("reason").and_then(|r| r.as_str()) != Some("compiler-artifact") {
-            continue;
-        }
-        let target = msg.get("target");
-        let target_name = target.and_then(|t| t.get("name")).and_then(|n| n.as_str());
-        let name_matches = matches!(target_name, Some(t) if t == cargo_name || t == normalized);
-        if !name_matches {
-            continue;
-        }
-        let is_cdylib = target
-            .and_then(|t| t.get("kind"))
-            .and_then(|k| k.as_array())
-            .map(|arr| arr.iter().any(|v| v.as_str() == Some("cdylib")))
-            .unwrap_or(false);
-        if !is_cdylib {
-            continue;
-        }
-        let filenames = msg
-            .get("filenames")
-            .and_then(|f| f.as_array())
-            .map(|arr| arr.iter().filter_map(|v| v.as_str()).collect::<Vec<_>>())
-            .unwrap_or_default();
-        for filename in filenames {
-            if filename.ends_with(&dot_ext) {
-                return Ok(Some(PathBuf::from(filename)));
-            }
-        }
-    }
-    Ok(None)
-}
-
 /// Reject `patch:` entries with a `path:` flavor at pack time. The
 /// resulting error names every offending entry so the dev can fix the
 /// manifest in one pass.
@@ -656,156 +458,6 @@ dependencies:
         );
         reject_path_patches_for_pack(dir.path())
             .expect("yaml without `patch:` block must pack cleanly");
-    }
-
-    #[test]
-    fn host_dylib_extension_matches_target_os() {
-        let ext = host_dylib_extension();
-        #[cfg(target_os = "macos")]
-        assert_eq!(ext, "dylib");
-        #[cfg(target_os = "windows")]
-        assert_eq!(ext, "dll");
-        #[cfg(all(not(target_os = "macos"), not(target_os = "windows")))]
-        assert_eq!(ext, "so");
-    }
-
-    #[test]
-    fn collect_host_dylibs_in_lib_returns_empty_when_dir_missing() {
-        let dir = tempdir().unwrap();
-        let lib = dir.path().join("lib");
-        let found = collect_host_dylibs_in_lib(&lib, "so").unwrap();
-        assert!(
-            found.is_empty(),
-            "missing lib/ dir should produce empty list, got: {:?}",
-            found
-        );
-    }
-
-    #[test]
-    fn collect_host_dylibs_in_lib_filters_by_extension() {
-        // Populate lib/ with one host-OS dylib and one non-matching file;
-        // the helper must pick the host file and skip the rest. Mentally
-        // reverting the extension filter would slurp every file in lib/
-        // and ship junk inside the slpkg.
-        let dir = tempdir().unwrap();
-        let lib = dir.path().join("lib");
-        std::fs::create_dir(&lib).unwrap();
-        std::fs::write(lib.join("libfoo.so"), b"so-bytes").unwrap();
-        std::fs::write(lib.join("libfoo.dylib"), b"dylib-bytes").unwrap();
-        std::fs::write(lib.join("README.md"), b"docs").unwrap();
-
-        let so_only = collect_host_dylibs_in_lib(&lib, "so").unwrap();
-        assert_eq!(so_only.len(), 1);
-        assert!(so_only[0].ends_with("libfoo.so"));
-
-        let dylib_only = collect_host_dylibs_in_lib(&lib, "dylib").unwrap();
-        assert_eq!(dylib_only.len(), 1);
-        assert!(dylib_only[0].ends_with("libfoo.dylib"));
-    }
-
-    #[test]
-    fn read_cargo_package_name_extracts_name_from_cargo_toml() {
-        // Mentally reverting the toml parse to a string-grep would happen
-        // to pass this case but break when a `[dependencies]` block carries
-        // a `name = "..."` line — the parser is the contract.
-        let dir = tempdir().unwrap();
-        write_cargo_toml(dir.path(), "grayscale-plugin");
-        let name = read_cargo_package_name(dir.path()).unwrap();
-        assert_eq!(name, "grayscale-plugin");
-    }
-
-    #[test]
-    fn read_cargo_package_name_errors_without_package_section() {
-        let dir = tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("Cargo.toml"),
-            r#"
-[workspace]
-members = ["foo"]
-"#,
-        )
-        .unwrap();
-        let err = read_cargo_package_name(dir.path())
-            .expect_err("workspace-only Cargo.toml must error");
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("[package].name"),
-            "error must point at the missing field, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn read_cargo_package_name_errors_when_cargo_toml_missing() {
-        // The error message must point at Cargo.toml so the user knows
-        // where to add it — the issue body's reference flow expects a
-        // Cargo.toml co-located with streamlib.yaml.
-        let dir = tempdir().unwrap();
-        let err = read_cargo_package_name(dir.path())
-            .expect_err("missing Cargo.toml must error");
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("Cargo.toml"),
-            "error must name Cargo.toml, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn parse_cargo_artifact_for_cdylib_returns_matching_host_dylib() {
-        // Real cargo-output shape sampled from
-        // `cargo build --release -p grayscale-plugin --message-format=json`
-        // against examples/camera-rust-plugin/plugin/: Cargo normalizes
-        // dashes-to-underscores in target.name, so a package
-        // `grayscale-plugin` emits target name `grayscale_plugin`. The
-        // filter has to accept BOTH spellings — match against just the
-        // dashed form (or just the underscore form) would silently fail.
-        // Reverting the kind / name / extension checks would pick the
-        // wrong crate or a non-cdylib file (rlib/intermediate).
-        let json = r#"
-{"reason":"compiler-artifact","target":{"name":"other-crate","kind":["lib"]},"filenames":["/tmp/target/release/libother.rlib"]}
-{"reason":"compiler-artifact","target":{"name":"grayscale_plugin","kind":["cdylib"]},"filenames":["/tmp/target/release/libgrayscale_plugin.so","/tmp/target/release/libgrayscale_plugin.d"]}
-{"reason":"build-finished","success":true}
-"#;
-        let found =
-            parse_cargo_artifact_for_cdylib(json, "grayscale-plugin", "so").unwrap();
-        assert_eq!(
-            found,
-            Some(PathBuf::from("/tmp/target/release/libgrayscale_plugin.so"))
-        );
-    }
-
-    #[test]
-    fn parse_cargo_artifact_for_cdylib_matches_dash_form_target_name() {
-        // A crate whose [lib].name explicitly retains a dash (rare but
-        // legal — Cargo allows it via `[lib].name = "foo-bar"`) emits a
-        // dash-form target.name. The filter must accept that too.
-        let json = r#"
-{"reason":"compiler-artifact","target":{"name":"grayscale-plugin","kind":["cdylib"]},"filenames":["/tmp/libgrayscale_plugin.so"]}
-"#;
-        let found =
-            parse_cargo_artifact_for_cdylib(json, "grayscale-plugin", "so").unwrap();
-        assert_eq!(found, Some(PathBuf::from("/tmp/libgrayscale_plugin.so")));
-    }
-
-    #[test]
-    fn parse_cargo_artifact_for_cdylib_ignores_unrelated_crate_artifacts() {
-        let json = r#"
-{"reason":"compiler-artifact","target":{"name":"some-other","kind":["cdylib"]},"filenames":["/tmp/libother.so"]}
-"#;
-        let found =
-            parse_cargo_artifact_for_cdylib(json, "grayscale-plugin", "so").unwrap();
-        assert!(found.is_none());
-    }
-
-    #[test]
-    fn parse_cargo_artifact_for_cdylib_returns_none_when_no_cdylib_built() {
-        // rlib-only build: no cdylib should be picked even though the
-        // crate name matches.
-        let json = r#"
-{"reason":"compiler-artifact","target":{"name":"grayscale-plugin","kind":["lib"]},"filenames":["/tmp/libgrayscale_plugin.rlib"]}
-"#;
-        let found =
-            parse_cargo_artifact_for_cdylib(json, "grayscale-plugin", "so").unwrap();
-        assert!(found.is_none());
     }
 
     #[test]
